fact_querying/requirements.txt
https://github.com/eCAPTION/ecaption_utils/archive/refs/heads/main.zip
python-dotenv
